Warehouse monitoring/raspberry-pi/requirements.txt
paho-mqtt==1.6.1
Adafruit-DHT==1.4.0
gpiozero==1.6.2
